backend/requirements.txt
amqp==2.5.1
autopep8==1.4.4
billiard==3.6.1.0
celery==4.3.0
Django==2.2.4
django-environ==0.4.5
django-extensions==2.1.9
django-js-asset==1.2.2
django-mptt==0.10.0
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
Pillow==6.1.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-magic==0.4.15
pytz==2019.2
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
